requirements.txt
python-dotenv
google-generativeai
grpcio
pydantic
